Removed hack in simple_type_witness implementation
the hack was to declare a struct to do a type equality bound
with all type arguments of the enum, without the `__Wit` type parameter.

This fix allows using `__Wit` in bounds, which was not possible before.

Added test for all permutations of trailing/not-trailing commas,
for simple_type_witness macro.

Added private macro to collect tokens of generic arguments

Updated changelog